Split off tag attributes code
